Metadata for rewrite rules.
Can now store if a rule is equational, and whether it is derived.
Equational metadata will be used to check the direction of rewrites in a derivation.
Derived metadata will be used to determine which rules require semantic checks.
In the future, equational metadata will be used to automate proof search, etc.