[Chore] Improved cases of error propogation code.
The error propogation functions require that all errors that require require propogation correction have cases.
The current code uses an otherwise case to catch all errors that do not require a correction.
The code would still compile if a new error was added, but the appropriate propogation code was not updated.
The new version of the code will fail to compile if a new error is not handled explicitly.
Comments have been added to justify this design decision.