Added framework to test executable logic.
For each test, the framework creates a temporary file to write all output into.
After the test, the file is read back, and the data is passed to a validation function.
The validation function determines the outcome of the test.

The framework has been applied to CheckGenerators to ensure that the framework works.