Derived field of RewriteRule now specifies source.
Originally we only cared if a rule was derived. However, this is not sufficient for cycle detection.
To explore the dependency graph of a proof, it must be possible to associate each derived rule with another proof.
This means that the 'derived' field should be a 'derivedFrom' field, whose data is a unique identifier.
Since each proof exposing a derived rule has a unique name, then it is sufficient to use the name as a unique identifier.