Cycle detection uses AbsRewrites.
This is needed to support equational derived rules.
This is because checking equationality requires that the dependency graph is acyclic.
Until equatioanlity is checked, all derived relations are abstracted by apply commands.

This has the added benefit of simplifying the cycle detection code (in terms of correctness).
Now there is a strongly typed difference between primitive and derived relations.
We no longer rely on metadata.